[Refactor] #4025 rooms-pit-nest.cpp/h からrooms-nest.cpp/h を分離した
