+++ Hengband/Hengband/Hengband.vcxproj.filters
@@ -2496,6 +2496,9 @@
     <ClCompile Include="..\..\src\room\rooms-pit.cpp">
       <Filter>room</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\room\rooms-nest.cpp">
+      <Filter>room</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\combat\shoot.h">
@@ -5397,8 +5400,13 @@
     <ClInclude Include="..\..\src\monster-race\race-misc-flags.h">
       <Filter>monster-race</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\monster-race\race-special-flags.h" />
+    <ClInclude Include="..\..\src\monster-race\race-special-flags.h">
+      <Filter>monster-race</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\room\rooms-pit.h">
+      <Filter>room</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\room\rooms-nest.h">
       <Filter>room</Filter>
     </ClInclude>
   </ItemGroup>
