+++ src/Makefile.am
@@ -786,6 +786,7 @@ hengband_SOURCES = \
 	room/rooms-city.cpp room/rooms-city.h \
 	room/rooms-fractal.cpp room/rooms-fractal.h \
 	room/rooms-maze-vault.cpp room/rooms-maze-vault.h \
+	room/rooms-nest.cpp room/rooms-nest.h \
 	room/rooms-normal.cpp room/rooms-normal.h \
 	room/rooms-pit.cpp room/rooms-pit.h\
 	room/rooms-pit-nest.cpp room/rooms-pit-nest.h \
