+++ src/room/room-generator.cpp
@@ -7,6 +7,7 @@
 #include "room/room-types.h"
 #include "room/rooms-city.h"
 #include "room/rooms-fractal.h"
+#include "room/rooms-nest.h"
 #include "room/rooms-normal.h"
 #include "room/rooms-pit-nest.h"
 #include "room/rooms-pit.h"
