+++ src/room/rooms-nest.h
@@ -1,0 +1,5 @@
+#pragma once
+
+class PlayerType;
+struct dun_data_type;
+bool build_type5(PlayerType *player_ptr, dun_data_type *dd_ptr);
