+++ src/room/rooms-pit-nest.h
@@ -50,7 +50,5 @@ struct nest_mon_info_type {
     bool used = false; //!< 既に選んだかどうか
 };
 
-struct dun_data_type;
-std::string pit_subtype_string(int type, bool nest);
 int pick_vault_type(const std::vector<nest_pit_type> &np_types, uint16_t allow_flag_mask, int dun_level);
-bool build_type5(PlayerType *player_ptr, dun_data_type *dd_ptr);
+std::string pit_subtype_string(int type, bool nest);
